Updated Breadcrumb to Prevent Chaining Render
Up to this point, the breadcrumb has not been refreshing to correctly match the
url. As a consequence, links would quickly concatonate, one after the other, in
the following fashion: Student > Student > Student > Student > 1.

To stop this, I simply pulled the JSX into a constant that will
update whenever the router changes. Although you will still be greeted by
'Student > Student' if you click on the 'My Students' tab, this is now a
consquence of the routing arrangement, which needs needs some TLC of its own.

Going forward, we will likely want to render Proper nouns such as School and
Student names in the Breadcrumb instead of primary keys. For all I know, that
was the original purpose of the Breadcrumb issue- unfortunately I was not
present to hear the why of its posting. Because we will want continue using
primary keys in the url, a solution would likely entail accessing state from
within the BreadCrumb component to swap the primary key with it's correspondant
Proper Noun whenever a number is encountered in the url.

Sounds like a Brenden job!